Modify view for profile desa (finish) and editing function for update profile desa

+++ database/migrations/2024_09_04_144215_create_realisasi_anggarans_table.php
@@ -16,17 +16,13 @@ class CreateRealisasiAnggaransTable extends Migration
         Schema::create('realisasi_anggarans', function (Blueprint $table) {
             $table->increments('realisasi_id');
 
-            // Menggunakan unsignedInteger untuk foreign key, bukan increments
-            $table->unsignedInteger('desa_id');
+            // Ubah ke unsignedBigInteger
+            $table->unsignedBigInteger('desa_id');
 
             $table->date('tahun');
             $table->integer('belanja_realisasi');
             $table->integer('dana_tidak_terpakai');
-
-            // Menggunakan string daripada char untuk fleksibilitas
             $table->string('laporan', 255);
-            
-            // Mendukung soft delete
             $table->softDeletes();
 
             // Tambahkan foreign key untuk desa_id
@@ -41,7 +37,6 @@ class CreateRealisasiAnggaransTable extends Migration
      */
     public function down()
     {
-        // Menggunakan dropIfExists untuk keamanan
         Schema::dropIfExists('realisasi_anggarans');
     }
 }
